repository layer

+++ app/src/main/java/com/example/fitnessapp/Data/Model/Entities/ActiveRun.kt
@@ -5,7 +5,7 @@ import com.example.fitnessapp.Data.Model.LocationPoints
 data class ActiveRun(
     val startTime : Long,
     val elapsedTime : Long,
-    val currentDistance : Double,
+    val currentDistance : Float,
     val route : List<LocationPoints>,
     val trackingStatus : Boolean = false
 )
